# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hospital API",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica un usuario.",
        "description": "Valida las credenciales del usuario y devuelve un token JWT\r\nque deberá enviarse en el encabezado Authorization para acceder\r\na los endpoints protegidos.\r\n            \r\nEjemplo:\r\n            \r\nPOST /api/Auth/login\r\n            \r\n{\r\n  \"username\": \"admin\",\r\n  \"password\": \"123456\"\r\n}",
        "requestBody": {
          "description": "Credenciales del usuario.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticación exitosa."
          },
          "401": {
            "description": "Credenciales inválidas.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Crea una nueva orden.",
        "description": "Registra una orden para un paciente indicando el servicio solicitado\r\ny la prioridad.",
        "parameters": [
          {
            "name": "PatientId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ServiceCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ServiceDescription",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Priority",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderPriority"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Orden creada correctamente."
          },
          "400": {
            "description": "Los datos enviados son inválidos.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Consulta las órdenes registradas.",
        "description": "Permite filtrar por paciente y por estado de la orden.\r\nAmbos parámetros son opcionales.",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "description": "Identificador del paciente.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Estado de la orden.",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta realizada correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Orders/{Id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Obtiene las ordenes de un paciente.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Identificador único del paciente.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": "La orden no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginCommand": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "nullable": true
          },
          "patientName": {
            "type": "string",
            "nullable": true
          },
          "serviceCode": {
            "type": "string",
            "nullable": true
          },
          "serviceDescription": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "$ref": "#/components/schemas/OrderPriority"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "processedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderPriority": {
        "enum": [
          "Normal",
          "Urgente"
        ],
        "type": "string"
      },
      "OrderStatus": {
        "enum": [
          "Pending",
          "Processed"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token JWT como: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}